Crate pallet_posts

Posts Module

Posts are the second crucial component of Subsocial after Spaces. This module allows you to create, update, move (between spaces), and hide posts as well as manage owner(s).

Posts can be compared to existing entities on web 2.0 platforms such as:

Posts on Facebook,

Tweets on Twitter,

Images on Instagram,

Articles on Medium,

Shared links on Reddit,

Questions and answers on Stack Overflow.

Modules

functions

rpc

weights
THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 2.0.1

Structs

Comment

Module

NextPostId
The next post id.

Post
Information about a post’s owner, its’ related space, content, and visibility.

PostById
Get the details of a post by its’ id.

PostIdsBySpaceId
Get the ids of all posts in a given space, by the space’s id.

PostUpdate

ReplyIdsByPostId
Get the ids of all direct replies by their parent’s post id.

SharedPostIdsByOriginalPostId
Get the ids of all posts that have shared a given original post id.

Enums

Call
Dispatchable calls.

Error

PostExtension
Post extension provides specific information necessary for different kinds of posts such as regular posts, comments, and shared posts.

RawEvent
Events for this module.

Constants

FIRST_POST_ID

Traits

AfterPostUpdated

PostScores

Trait
The pallet’s configuration trait.

WeightInfo

Type Definitions

Event
RawEvent specialized for the configuration Trait